+++ src/Rerefined.hs
@@ -1,0 +1,7 @@
+module Rerefined
+  ( module Rerefined.Refine
+  , Predicate, Refine, Refine1
+  ) where
+
+import Rerefined.Refine
+import Rerefined.Predicate