+++ src/Rerefined/Refined1.hs
@@ -1,10 +1,15 @@
+-- | 'Refined1' definition.
+--
+-- Not intended for external use. For unsafe refines, use
+-- 'Rerefined.Refine.Unsafe'.
+
 module Rerefined.Refined1 where
 
 import Language.Haskell.TH.Syntax ( Lift )
 
 -- | @f a@ refined with predicate @p@.
 newtype Refined1 p f a = Refined1 (f a)
-    deriving stock (Lift, Show) -- TODO Show? useful but meh?
+    deriving stock (Functor, Lift, Show) -- TODO Show? useful but meh?
 
 -- | Strip the refinement from a 'Refined1'.
 --