+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+* add missing `Show` instance to `RefineFailure`
+* add missing `Functor` instance to `Refined1`
+* general cleanup
+
 ## 0.1.0 (2024-04-30)
 Initial release.